Add FileValidations concern for illustration and blueprint models
issue #18
