+++ app/models/illustration.rb
@@ -15,28 +15,15 @@
 #
 
 class Illustration < ApplicationRecord
+  include FileValidations
+
   has_one :design_illustration, inverse_of: :illustration
   has_one :design, through: :design_illustration
 
   #
   # Validations
+  # FileValidations (concern)
   #
 
   validates_format_of :url_path, with: /\.(png|jpg|jpeg|gif)\z/i
-  # TODO: MOVE SHARED
-  validate :filename_is_empty
-  before_validation :sanitize_attrs
-
-  def filename_is_empty
-    # .png
-    errors.add(:filename, :empty) if File.basename(url_path, '.*').split('.')[0] == ''
-  end
-
-  def sanitize_attrs
-    # "><svg onload=alert(1)>.jpg
-    %i[url url_path].each do |attr|
-      value = self[attr]
-      self[attr] = Sanitize.fragment(value) if value.present?
-    end
-  end
 end
